tests/submission/MOSQUITO/search_input.html

<!DOCTYPE html>
<html>

  <head>
    <title>Search Input</title>
     <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<meta name="author" lang="fr" content="Fabrice Clari, Dimitri Bocquet">
   <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  </head>

  <body>

  
  		<h1>Search Input</h1>
  		<input type="search" style="display:none" placeholder="Search..." />

    <div id="log">
    </div>
	
	<script type="text/javascript">
		

		test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "search")}, "search type support on input element");
		test(function() {assert_equals(document.getElementsByTagName("input")[0].placeholder, "Search...")}, "placeholder attribute support on input element");
		  
	</script>
	
  </body>
  
</html>
